+++ README.md
@@ -8,12 +8,16 @@ buttondown.email + react = 💙
 yarn add react-buttondown
 ```
 
-### Usage
+### API
 
-`react-buttondown` exports, at the moment, a single `Buttondown` component that serves as a pre-built form for capturing subscribers.
+`react-buttondown` exports a drop-in form and a custom hook that exposes [buttondown's api functionality](https://api.buttondown.email/v1/schema).
+
+#### `ButtondownForm`
+
+A drop-in form that captures subscriptions.
 
 ```js
-import { Buttondown } from 'react-buttondown'
+import { ButtondownForm } from 'react-buttondown'
 import 'react-buttondown/dist/react-buttondown.css'
 
 function App() {
@@ -21,7 +25,7 @@ function App() {
   const handleOnSubscribe = subscriber => { /* ... */ } 
 
   return (
-    <Buttondown 
+    <ButtondownForm 
       apiKey={API_KEY} 
       onSubscribe={handleOnSubscribe} 
     />
@@ -38,6 +42,48 @@ function App() {
 `onSubscribe` (_optional_)
 
 A function handler that receives the subscriber object after successfully subscribing a user.
+
+#### `useButtondown`
+
+A custom hook that exposes [buttondown's api functionality](https://api.buttondown.email/v1/schema).
+
+```js
+import { useButtondown } from 'react-buttondown'
+
+function App() {
+  const { addSubscriber } = useButtondown('my-api-key')
+
+  const handleAddSubscriber = async (email) => {
+	try {
+      const response = await addSubscriber(email)
+      const subscriber = response.data
+
+    } catch (err) {
+      const errorMessage = err.data
+      console.log(errorMessage)
+	}
+  } 
+
+// ...
+}
+```
+
+**arguments**:
+
+`apiKey`
+
+- Your [buttondown.email](https://buttondown.email/) api key. This can be found in your [newsletter settings](https://buttondown.email/settings).
+
+
+**returns**:
+
+`addSubscriber` 
+
+- An async function that takes a string `email` argument and returns an object with returned data and the response status:
+
+```js
+const { data, response } = await addSubscriber(email)
+```
 
 ### Contributing
